# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistem Kasir Enterprise API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PRODUCTS
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers/members
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer/member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/transactions:
    get:
      operationId: getCustomerTransactions
      tags: [customers]
      summary: Get transactions for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # SHIFTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # USER
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, admin, cashier, warehouse, accountant]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - name
        - role
        - isActive
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        productCount:
          type: integer
      required:
        - id
        - name
        - productCount
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
      required:
        - name

    # PRODUCT
...
    TransactionItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        price:
          type: number
        discount:
          type: number
        subtotal:
          type: number
      required:
        - id
        - productId
        - productName
        - sku
        - quantity
        - price
        - discount
        - subtotal
    TransactionPayment:
      type: object
      properties:
        method:
          type: string
          enum: [cash, card, qris, transfer, voucher, points]
        amount:
          type: number
      required:
        - method
        - amount
    Transaction:
      type: object
      properties:
        id:
          type: integer
        receiptNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        cashierId:
          type: integer
        cashierName:
          type: string
        shiftId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransactionItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/TransactionPayment"
        subtotal:
          type: number
        discountAmount:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        paidAmount:
          type: number
        changeAmount:
          type: number
        status:
          type: string
          enum: [completed, voided, returned]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - receiptNumber
        - cashierId
        - cashierName
        - items
        - payments
        - subtotal
        - discountAmount
        - taxAmount
        - total
        - paidAmount
        - changeAmount
        - status
        - createdAt
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        memberCode:
          type: ["string", "null"]
        tier:
          type: string
          enum: [regular, silver, gold, platinum]
        points:
          type: integer
        totalSpent:
          type: number
        totalTransactions:
          type: integer
        birthdate:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - tier
        - points
        - totalSpent
        - totalTransactions
        - isActive
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
      required:
        - name
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
        isActive:
          type: boolean
    CustomerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # SHIFTS
...